# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Medical Appointment API
  version: 1.0.0
  description: API documentation for Medical Appointment System
servers:
  - url: http://localhost:5000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - name
        - email
        - password
        - phoneNumber
        - CIN
        - role
      properties:
        name:
          type: string
          example: "Dr. John Smith"
        email:
          type: string
          format: email
          example: "john.smith@example.com"
        password:
          type: string
          format: password
          example: "Doctor123!"
        phoneNumber:
          type: string
          example: "+1234567890"
        CIN:
          type: string
          example: "D123456"
        role:
          type: string
          enum: [Doctor, Patient, Admin]
        specialization:
          type: string
          example: "Cardiology"
        department:
          type: string
          format: uuid
        profileImage:
          type: string
          format: uri
        diplomaImage:
          type: string
          format: uri
        availability:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
                enum: [Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, Sunday]
              slots:
                type: array
                items:
                  type: object
                  properties:
                    time:
                      type: string
                      example: "09:00"
                    isBooked:
                      type: boolean
                      default: false
        isValidated:
          type: boolean
          default: false
        medicalHistory:
          type: string
...
    MedicalHistory:
      type: object
      required:
        - patient
        - doctor
        - diagnosis
      properties:
        patient:
          type: string
          format: uuid
        doctor:
          type: string
          format: uuid
        appointment:
          type: string
          format: uuid
        diagnosis:
          type: string
        prescription:
          type: array
          items:
            type: object
            properties:
              medication:
                type: string
              dosage:
                type: string
              frequency:
                type: string
              duration:
                type: string
        notes:
          type: string
        attachments:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              description:
                type: string
        vitalSigns:
          type: object
          properties:
            bloodPressure:
              type: string
            heartRate:
              type: string
            temperature:
              type: string
        followUpDate:
          type: string
          format: date-time
paths:
...
  /api/doctors/{id}/patients:
    get:
      tags: [Doctors]
      summary: Get doctor's patients
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        200:
          description: List of doctor's patients
...
  /api/patients:
    get:
      tags: [Patients]
      summary: Get all patients
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of all patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /api/patients/profile:
    get:
      tags: [Patients]
      summary: Get patient profile
      security:
        - bearerAuth: []
      responses:
        200:
          description: Patient profile
    
    put:
      tags: [Patients]
      summary: Update patient profile
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phoneNumber:
                  type: string
                email:
                  type: string
      responses:
        200:
          description: Profile updated successfully
  /api/patients/my-appointments:
    get:
      tags: [Patients]
      summary: Get patient's appointments
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of patient's appointments
  /api/patients/my-medical-history:
    get:
      tags: [Patients]
      summary: Get patient's medical history
      security:
        - bearerAuth: []
      responses:
        200:
          description: Patient's medical history
  /api/patients/book-appointment:
    post:
      tags: [Patients]
      summary: Book an appointment
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - doctorId
                - appointmentDate
                - reason
                - departmentId
              properties:
                doctorId:
                  type: string
                appointmentDate:
                  type: string
                  format: date-time
                reason:
                  type: string
                departmentId:
                  type: string
      responses:
        201:
          description: Appointment booked successfully
...
  /api/admin/stats:
    get:
      tags: [Admin]
      summary: Get overall statistics
      security:
        - bearerAuth: []
      responses:
        200:
          description: Overall system statistics
  /api/admin/verify-doctor/{id}:
    put:
      tags: [Admin]
      summary: Verify doctor account
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isValidated:
                  type: boolean
      responses:
        200:
          description: Doctor verification status updated successfully
...
  /api/medical-history/doctor/patients:
    get:
      tags: [Medical History]
      summary: Get all medical histories for doctor's patients
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of medical histories grouped by patient
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    patient:
                      type: object
                      properties:
                        _id:
                          type: string
                        name:
                          type: string
                        email:
                          type: string
                        phoneNumber:
                          type: string
                        CIN:
                          type: string
                    histories:
                      type: array
                      items:
                        $ref: '#/components/schemas/MedicalHistory'
        403:
          description: Access denied. Doctors only.
        500:
          description: Server error
security:
  - bearerAuth: [] 